requirements.txt
fastapi
uvicorn[standard]
pydantic
sqlalchemy
pinecone
google-generativeai
python-dotenv
pypdf
python-docx
python-multipart
requests
httpx  